refactor: eval admin schema
