+++ src/controllers/admin/evaluation-administration-controller.ts
@@ -5,7 +5,6 @@ import { ValidationError } from "yup"
 import * as EvaluationAdministrationService from "../../services/evaluation-administration-service"
 import * as EvaluationResultService from "../../services/evaluation-result-service"
 
-import { createEvaluationSchema } from "../../utils/validation/evaluations/create-evaluation-schema"
 import prisma from "../../utils/prisma"
 import { EvaluationAdministrationStatus } from "../../types/evaluation-administration-type"
 import { EvaluationStatus } from "../../types/evaluation-type"
@@ -15,6 +14,7 @@ import CustomError from "../../utils/custom-error"
 import {
   addEvaluatorSchema,
   addExternalEvaluatorsSchema,
+  createEvaluationAdministrationSchema,
 } from "../../utils/validation/evaluation-administration-schema"
 
 /**
@@ -61,7 +61,7 @@ export const store = async (req: Request, res: Response) => {
       email_content,
     } = req.body
 
-    await createEvaluationSchema.validate({
+    await createEvaluationAdministrationSchema.validate({
       name,
       eval_period_start_date,
       eval_period_end_date,
@@ -134,7 +134,7 @@ export const update = async (req: Request, res: Response) => {
       email_content,
     } = req.body
 
-    await createEvaluationSchema.validate({
+    await createEvaluationAdministrationSchema.validate({
       name,
       eval_period_start_date,
       eval_period_end_date,
